Change TextConfig.width to f32
- Change `TextConfig.width` from `int` to `f32` in [layout_types.v](cci:7://file:///Users/mike/Documents/github/vglyph/layout_types.v:0:0-0:0)
- Update [layout.v](cci:7://file:///Users/mike/Documents/github/vglyph/layout.v:0:0-0:0) to handle `f32` width
- Update API documentation
- Add `examples/size_test_demo.v` to `_check.vsh`

+++ layout.v
@@ -169,7 +169,7 @@ fn setup_pango_layout(mut ctx Context, text string, cfg TextConfig) !&C.PangoLay
 
 	// Apply layout configuration
 	if cfg.width > 0 {
-		C.pango_layout_set_width(layout, cfg.width * pango_scale)
+		C.pango_layout_set_width(layout, int(cfg.width * pango_scale))
 		pango_wrap := match cfg.wrap {
 			.word { PangoWrapMode.pango_wrap_word }
 			.char { PangoWrapMode.pango_wrap_char }
